requirements.txt
# cairocffi==1.1.0
# certifi==2020.6.20
# cffi==1.14.3
# cycler==0.10.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
grpcio-tools==1.32.0
# kiwisolver==1.2.0
lxml==4.6.1
matplotlib==3.3.2
numpy==1.19.2
pandas==1.1.3
Pillow==8.0.0
protobuf==3.13.0
# pycparser==2.20
# pyparsing==2.4.7
# python-dateutil==2.8.1
# pytz==2020.1
# qtile==0.16.1
scipy==1.5.3
# six==1.15.0
# wakatime==13.0.7
# xcffib==0.10.1
